+++ wagtail/wagtailadmin/templates/wagtailadmin/pages/_editor_js.html
@@ -21,8 +21,6 @@
     <script src="{% static 'admin/js/urlify.js' %}"></script>
     <script src="{% static 'wagtailadmin/js/privacy-switch.js' %}"></script>
 
-    {% include "wagtailrichtexteditor/pages/_richtexteditor_js.html" %}
-
     {% hook_output 'insert_editor_js' %}
 {% endcompress %}
