+++ wagtail/wagtailembeds/static_src/wagtailembeds/js/hallo-plugins/hallo-wagtailembeds.js
