+++ wagtail/wagtailhallo/static_src/wagtailhallo/js/hallo-plugins/hallo-wagtaillink.js
