+++ wagtail/wagtailhallo/static_src/wagtailhallo/js/rich-text-editor.js
