+++ wagtail/wagtailimages/static_src/wagtailimages/js/hallo-plugins/hallo-wagtailimage.js
