+++ wagtail/wagtailimages/wagtail_hooks.py
@@ -49,6 +49,23 @@ def editor_js():
     )
 
 
+@hooks.register('insert_hallo_js')
+def images_richtexteditor_js():
+    js_files = [
+        'wagtailimages/js/hallo-plugins/hallo-wagtailimage.js',
+    ]
+    js_includes = format_html_join('\n', '<script src="{0}{1}"></script>',
+        ((settings.STATIC_URL, filename) for filename in js_files)
+    )
+    return js_includes + format_html(
+        """
+        <script>
+            registerHalloPlugin('hallowagtailimage');
+        </script>
+        """,
+    )
+
+
 @hooks.register('register_permissions')
 def register_permissions():
     return Permission.objects.filter(content_type__app_label='wagtailimages',
